# make `fastlog2` faster