This uses the same formula, but slightly rearranged. The primary difference is that the coefficients have been shifted slightly such that subtracting 1 from `signif` isn't necessary to computer `quot`  which removes instruction dependencies and leads to slightly faster code. I've also unified the code paths leading to slightly simpler code which GCC apparently tripped up on, but LLVM seems fine with.

```
julia> @btime fastlog2(x) setup = (x=rand(Float32));
  3.270 ns (0 allocations: 0 bytes) #before
  2.730 ns (0 allocations: 0 bytes) #after 

julia> @btime fastpower(x,y) setup = (x,y)=rand(Float32),rand(Float32);
14.850 ns (0 allocations: 0 bytes) #before
  8.929 ns (0 allocations: 0 bytes) #after
```